enh: externalize blog functions WIP
